requirements.txt
hdbcli
python-dotenv
requests
tzdata
